Move PanCacheData and related functions to pan-view
Rename pan_item_image_find_size() to pan_cache_get_image_size() and
simplify a bit.
