+++ src/pan-view/pan-view.h
@@ -26,6 +26,7 @@
 
 #include "typedefs.h"
 
+class FileData;
 struct PanItem;
 struct PanWindow;
 
@@ -33,9 +34,9 @@ void pan_layout_update(PanWindow *pw);
 GList *pan_layout_intersect(PanWindow *pw, gint x, gint y, gint width, gint height);
 void pan_layout_resize(PanWindow *pw);
 
-void pan_cache_sync_date(PanWindow *pw, GList *list);
-
 GList *pan_cache_sort(GList *list, SortType method, gboolean ascend, gboolean case_sensitive);
+void pan_cache_sync_date(PanWindow *pw, GList *list);
+void pan_cache_get_image_size(PanWindow *pw, const FileData *fd, gint &w, gint &h);
 
 void pan_info_update(PanWindow *pw, PanItem *pi);
